chore: interface 계층의 패키지 구조 개편
- 패키지 하위에 도메인 패키지들이 위치하도록 변경
