+++ src/main/kotlin/kr/hhplus/be/server/interfaces/balance/BalanceRequest.kt
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.interfaces.balance
+
+sealed class BalanceRequest {
+    data class Charge(val amount: Int) : BalanceRequest()
+}
