+++ src/main/kotlin/kr/hhplus/be/server/interfaces/coupon/CouponController.kt
@@ -1,8 +1,5 @@
-package kr.hhplus.be.server.interfaces.controller
+package kr.hhplus.be.server.interfaces.coupon
 
-import kr.hhplus.be.server.interfaces.api.CouponApi
-import kr.hhplus.be.server.interfaces.dto.request.CouponIssueRequest
-import kr.hhplus.be.server.interfaces.dto.response.CouponIssueResponse
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
@@ -10,8 +7,8 @@ import java.time.Instant
 
 @RestController
 class CouponController : CouponApi {
-    override fun issue(id: Long, @RequestBody request: CouponIssueRequest): ResponseEntity<CouponIssueResponse> {
-        val response = CouponIssueResponse(
+    override fun issue(id: Long, @RequestBody request: CouponRequest.Issue): ResponseEntity<CouponResponse.Issue> {
+        val response = CouponResponse.Issue(
             couponId = id,
             customerId = request.customerId,
             status = "ISSUED",
