+++ src/main/kotlin/kr/hhplus/be/server/interfaces/coupon/CouponRequest.kt
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.interfaces.coupon
+
+sealed class CouponRequest {
+    data class Issue(val customerId: Long) : CouponRequest()
+}
