+++ src/main/kotlin/kr/hhplus/be/server/interfaces/customer/CustomerController.kt
@@ -1,11 +1,9 @@
-package kr.hhplus.be.server.interfaces.controller
+package kr.hhplus.be.server.interfaces.customer
 
 import kr.hhplus.be.server.application.balance.BalanceFacade
-import kr.hhplus.be.server.interfaces.api.CustomerApi
-import kr.hhplus.be.server.interfaces.dto.request.BalanceChargeRequest
-import kr.hhplus.be.server.interfaces.dto.response.BalanceHistoryResponse
-import kr.hhplus.be.server.interfaces.dto.response.BalanceResponse
-import kr.hhplus.be.server.interfaces.dto.response.CustomerCouponResponse
+import kr.hhplus.be.server.interfaces.balance.BalanceRequest
+import kr.hhplus.be.server.interfaces.balance.BalanceResponse
+import kr.hhplus.be.server.interfaces.coupon.CouponResponse
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
@@ -14,14 +12,14 @@ import org.springframework.web.bind.annotation.RestController
 class CustomerController(
     private val balanceFacade: BalanceFacade
 ) : CustomerApi {
-    override fun getBalance(id: Long): ResponseEntity<BalanceResponse> {
+    override fun getBalance(id: Long): ResponseEntity<BalanceResponse.Summary> {
         val balance = balanceFacade.getByCustomerId(id)
         return ResponseEntity.ok(BalanceResponse.from(balance))
     }
 
-    override fun getBalanceHistories(id: Long): ResponseEntity<List<BalanceHistoryResponse>> {
+    override fun getBalanceHistories(id: Long): ResponseEntity<List<BalanceResponse.Histories>> {
         val response = listOf(
-            BalanceHistoryResponse(
+            BalanceResponse.Histories(
                 changeType = "CHARGE",
                 changeAmount = 10000,
                 totalAmount = 60000,
@@ -32,13 +30,16 @@ class CustomerController(
         return ResponseEntity.ok(response)
     }
 
-    override fun chargeBalance(id: Long, @RequestBody request: BalanceChargeRequest): ResponseEntity<BalanceResponse> {
-        return ResponseEntity.ok(BalanceResponse(id, 150000))
+    override fun chargeBalance(
+        id: Long,
+        @RequestBody request: BalanceRequest.Charge
+    ): ResponseEntity<BalanceResponse.Summary> {
+        return ResponseEntity.ok(BalanceResponse.Summary(id, 150000))
     }
 
-    override fun getCustomerCoupons(id: Long): ResponseEntity<List<CustomerCouponResponse>> {
+    override fun getCustomerCoupons(id: Long): ResponseEntity<List<CouponResponse.Owned>> {
         val response = listOf(
-            CustomerCouponResponse(
+            CouponResponse.Owned(
                 name = "첫 구매 할인",
                 discountType = "FIXED",
                 discountAmount = 3000,
@@ -46,7 +47,7 @@ class CustomerController(
                 issuedAt = "2025-04-02T15:00:00Z",
                 expiredAt = "2025-04-30T23:59:59Z"
             ),
-            CustomerCouponResponse(
+            CouponResponse.Owned(
                 name = "봄맞이 프로모션",
                 discountType = "PERCENT",
                 discountAmount = 10,
