--- src/main/kotlin/kr/hhplus/be/server/interfaces/dto/request/CouponIssueRequest.kt
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.interfaces.dto.request
-
-data class CouponIssueRequest(
-    val customerId: Long
-)
