--- src/main/kotlin/kr/hhplus/be/server/interfaces/dto/request/OrderItemRequest.kt
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.interfaces.dto.request
-
-data class OrderItemRequest(
-    val productId: Long,
-    val productOptionId: Long,
-    val quantity: Int
-)
