--- src/main/kotlin/kr/hhplus/be/server/interfaces/dto/request/OrderRequest.kt
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.interfaces.dto.request
-
-data class OrderRequest(
-    val customerId: Long,
-    val items: List<OrderItemRequest>
-)
