--- src/main/kotlin/kr/hhplus/be/server/interfaces/dto/response/BalanceHistoryResponse.kt
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.interfaces.dto.response
-
-data class BalanceHistoryResponse(
-    val changeType: String,
-    val changeAmount: Int,
-    val totalAmount: Int,
-    val createdAt: String
-)
