--- src/main/kotlin/kr/hhplus/be/server/interfaces/dto/response/OrderItemResponse.kt
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.interfaces.dto.response
-
-data class OrderItemResponse(
-    val productName: String,
-    val optionName: String,
-    val quantity: Int,
-    val subtotalPrice: Int
-)
