--- src/main/kotlin/kr/hhplus/be/server/interfaces/dto/response/OrderResponse.kt
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.interfaces.dto.response
-
-data class OrderResponse(
-    val orderId: Long,
-    val customerId: Long,
-    val totalPrice: Int,
-    val createdAt: String,
-    val items: List<OrderItemResponse>
-)
