--- src/main/kotlin/kr/hhplus/be/server/interfaces/dto/response/PopularProductResponse.kt
@@ -1,8 +1,0 @@
-package kr.hhplus.be.server.interfaces.dto.response
-
-data class PopularProductResponse(
-    val productId: Long,
-    val name: String,
-    val basePrice: Int,
-    val salesCount: Int
-)
