--- src/main/kotlin/kr/hhplus/be/server/interfaces/dto/response/ProductListResponse.kt
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.interfaces.dto.response
-
-data class ProductListResponse(
-    val id: Long,
-    val name: String,
-    val basePrice: Int
-)
