--- src/main/kotlin/kr/hhplus/be/server/interfaces/dto/response/ProductOptionResponse.kt
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.interfaces.dto.response
-
-data class ProductOptionResponse(
-    val optionName: String,
-    val extraPrice: Int
-)
