+++ src/main/kotlin/kr/hhplus/be/server/interfaces/order/OrderRequest.kt
@@ -1,0 +1,14 @@
+package kr.hhplus.be.server.interfaces.order
+
+sealed class OrderRequest {
+    data class Create(
+        val customerId: Long,
+        val items: List<OrderItem>
+    )
+
+    data class OrderItem(
+        val productId: Long,
+        val productOptionId: Long,
+        val quantity: Int
+    )
+}
