+++ src/main/kotlin/kr/hhplus/be/server/interfaces/payment/PaymentController.kt
@@ -1,8 +1,5 @@
-package kr.hhplus.be.server.interfaces.controller
+package kr.hhplus.be.server.interfaces.payment
 
-import kr.hhplus.be.server.interfaces.api.PaymentApi
-import kr.hhplus.be.server.interfaces.dto.request.PaymentRequest
-import kr.hhplus.be.server.interfaces.dto.response.PaymentResponse
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
