+++ src/main/kotlin/kr/hhplus/be/server/interfaces/payment/PaymentRequest.kt
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.dto.request
+package kr.hhplus.be.server.interfaces.payment
 
 data class PaymentRequest(
     val orderId: Long,
