+++ src/main/kotlin/kr/hhplus/be/server/interfaces/payment/PaymentResponse.kt
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.dto.response
+package kr.hhplus.be.server.interfaces.payment
 
 data class PaymentResponse(
     val paymentId: Long,
